minor refactos
